+++ packages/react-charts/rollup.config.mjs
@@ -1,0 +1,33 @@
+import esbuild from "rollup-plugin-esbuild";
+
+const outputDefaultConfig = {
+  preserveModules: true,
+  preserveModulesRoot: 'src',
+  sourcemap: false,
+};
+
+export default [
+  {
+    input: ['./src/index.ts'],
+    output: [
+      {
+        dir: 'dist/esm',
+        format: 'esm',
+        ...outputDefaultConfig,
+      },
+      {
+        dir: 'dist/cjs',
+        format: 'cjs',
+        exports: 'auto',
+        ...outputDefaultConfig,
+      },
+    ],
+    plugins: [esbuild({ tsconfig: 'tsconfig.build.json', target: 'esnext' })],
+    external: [
+      'react',
+      'react-native',
+      '@emotion/native',
+      '@emotion/react',
+    ],
+  },
+];