+++ packages/react-charts/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './svg';
+export * from './pie';
+export * from './font';