+++ packages/react-core/src/components/atoms/Avatar/index.ts
@@ -1,2 +1,2 @@
 export { default as Avatar } from './Avatar';
-export type { AvatarProps } from './Avatar';
+export type { AvatarProps } from './types';